GUI application

+++ helpers/pdf_helpers/delete_pdf.py
@@ -3,9 +3,10 @@ from tkinter import messagebox
 
 
 def delete_PDF():
-    file = "result.pdf"
-    try:
-        os.remove(file)
-    except PermissionError:
-        messagebox.showwarning('Warning', 'Close the file!')
-        delete_PDF()
+    if os.path.exists("result.pdf"):
+        try:
+            file = "result.pdf"
+            os.remove(file)
+        except PermissionError:
+            messagebox.showwarning('Warning', 'Close the file!')
+            delete_PDF()
